requirements.txt
requests
schedule
python_firebase
firebase
Werkzeug
Flask
Flask-RESTful
zappa
boto3
AWS
python_jwt
gcloud
sseclient
pycryptodome
ConfigParser
troposphere<3
requests_toolbelt
python-dotenv-1.0.0